+++ src/analyze/src/components/generic/MainNav.vue
@@ -26,6 +26,17 @@
         </router-link>
 
         <router-link
+          :to="{name: 'orchestration'}"
+          v-if="orchestrationEnabled"
+          class="navbar-item navbar-child">
+          Orchestration
+        </router-link>
+        <span v-else class="navbar-item navbar-child disabled">
+          <span data-tooltip="Airflow is required."
+                class="tooltip is-tooltip-right is-tooltip-bottom-desktop">Orchestration</span>
+        </span>
+
+        <router-link
           :to="{name: 'analyze'}"
           class="navbar-item navbar-child has-text-weight-semibold">
           Analyze
@@ -35,7 +46,6 @@
           class="navbar-item navbar-child has-text-weight-semibold">
           Dashboards
         </router-link>
-
       </div>
 
       <div class="navbar-end">
@@ -50,7 +60,6 @@
           </a>
         </div>
       </div>
-
     </div>
   </nav>
 </template>
@@ -73,6 +82,11 @@ export default {
     return {
       isMobileMenuOpen: false,
     };
+  },
+  computed: {
+    orchestrationEnabled() {
+      return Boolean(FLASK.airflowUrl);
+    },
   },
   methods: {
     mobileMenuClicked() {