+++ src/analyze/test/e2e/specs/test.js
@@ -3,7 +3,7 @@
 
 module.exports = {
   'default e2e tests': function test(browser) {
-    // automatically uses dev Server port from /config.index.js
+    // automatically uses dev Server port from /build/config.js
     // default: http://localhost:8080
     // see nightwatch.conf.js
     const devServer = browser.globals.devServerUrl;