+++ src/meltano/api/workers.py
@@ -9,7 +9,10 @@ from watchdog.observers import Observer
 from watchdog.events import PatternMatchingEventHandler, EVENT_TYPE_MODIFIED
 
 from meltano.core.project import Project
+from meltano.core.plugin import Plugin
+from meltano.core.config_service import ConfigService
 from meltano.core.compiler.project_compiler import ProjectCompiler
+from meltano.core.plugin_invoker import invoker_factory
 
 
 class CompileEventHandler(PatternMatchingEventHandler):
@@ -79,3 +82,23 @@ class UIAvailableWorker(threading.Thread):
 
     def stop(self):
         self._terminate = True
+
+
+class AirflowWorker(threading.Thread):
+    def __init__(self, project: Project, airflow: Plugin = None):
+        super().__init__()
+
+        self.project = project
+        self._plugin = airflow or ConfigService(project).get_plugin("airflow")
+
+    def start_all(self):
+        invoker = invoker_factory(self.project, self._plugin)
+        self._webserver = invoker.invoke("webserver")
+        self._scheduler = invoker.invoke("scheduler")
+
+    def run(self):
+        return self.start_all()
+
+    def stop(self):
+        self._webserver.terminate()
+        self._scheduler.terminate()